+++ src/components/DetailsPage.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+function DetailsPage() {
+  return (
+    <div>DetailsPage</div>
+  );
+}
+
+export default DetailsPage;
