--- src/pages/Details.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-function Detail() {
-  return (
-    <div>Detail</div>
-  );
-}
-
-export default Detail;
