--- src/pages/Layout.js
@@ -1,15 +1,0 @@
-import React from 'react';
-import { Outlet } from 'react-router-dom';
-import Navbar from '../components/Navbar';
-
-function Layout() {
-  return (
-    <>
-      <Navbar />
-      <hr />
-      <Outlet />
-    </>
-  );
-}
-
-export default Layout;
