+++ src/redux/home/homeSlice.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+function homeSlice() {
+  return (
+    <div>homeSlice</div>
+  );
+}
+
+export default homeSlice;
